Moving and adding more tests cases
