+++ Cake/Test/TestCase/ORM/ValidatorTest.php
@@ -349,4 +349,5 @@ class ValidatorTest extends \Cake\TestSuite\TestCase {
 
 		$this->assertEquals($expected, $errors);
 	}
+
 }
